Finish bag implementation, use vector instead of multiset, change piece values to be uint32_t, test in main

+++ include/Piece.h
@@ -6,14 +6,12 @@
 class Piece {
     private:
         char letter_;
-        std::uint8_t value_;
+        std::uint32_t value_;
         bool is_wildcard_;
     public: 
-        Piece(char letter, std::uint8_t val);
+        Piece(char letter, std::uint32_t val);
         void setLetter(char letter);
         char getLetter();
-        std::uint8_t getValue();
-
-
+        std::uint32_t getValue();
 };
 #endif
